+++ src/libServer/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_library(Server Server.cpp JSONConversion.cpp GetWorkServer.cpp)
+add_library(Server Server.cpp JSONConversion.cpp GetWorkServer.cpp LookupServer.cpp StatusServer.cpp)
 
 add_dependencies(Server jsonrpc-project)
 target_include_directories(Server PUBLIC ${PROJECT_SOURCE_DIR}/src ${JSONRPC_INCLUDE_DIR})
