+++ src/libServer/LookupServer.h
@@ -1,0 +1,242 @@
+/*
+ * Copyright (C) 2019 Zilliqa
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#ifndef __LOOKUP_SERVER_H__
+#define __LOOKUP_SERVER_H__
+
+#include "Server.h"
+
+class Mediator;
+
+class LookupServer : public Server,
+                     public jsonrpc::AbstractServer<LookupServer> {
+  std::mutex m_mutexBlockTxPair;
+  std::pair<uint64_t, boost::multiprecision::uint128_t> m_BlockTxPair;
+  std::mutex m_mutexTxBlockCountSumPair;
+  std::pair<uint64_t, boost::multiprecision::uint128_t> m_TxBlockCountSumPair;
+  uint64_t m_StartTimeTx;
+  uint64_t m_StartTimeDs;
+  std::mutex m_mutexDSBlockCache;
+  std::pair<uint64_t, CircularArray<std::string>> m_DSBlockCache;
+  std::mutex m_mutexTxBlockCache;
+  std::pair<uint64_t, CircularArray<std::string>> m_TxBlockCache;
+  static CircularArray<std::string> m_RecentTransactions;
+  static std::mutex m_mutexRecentTxns;
+  std::mt19937 m_eng;
+
+ public:
+  LookupServer(Mediator& mediator, jsonrpc::AbstractServerConnector& server);
+  ~LookupServer();
+
+  inline virtual void GetNetworkIdI(const Json::Value& request,
+                                    Json::Value& response) {
+    (void)request;
+    response = this->GetNetworkId();
+  }
+  inline virtual void CreateTransactionI(const Json::Value& request,
+                                         Json::Value& response) {
+    response = this->CreateTransaction(request[0u]);
+  }
+  inline virtual void GetTransactionI(const Json::Value& request,
+                                      Json::Value& response) {
+    response = this->GetTransaction(request[0u].asString());
+  }
+  inline virtual void GetDsBlockI(const Json::Value& request,
+                                  Json::Value& response) {
+    response = this->GetDsBlock(request[0u].asString());
+  }
+  inline virtual void GetTxBlockI(const Json::Value& request,
+                                  Json::Value& response) {
+    response = this->GetTxBlock(request[0u].asString());
+  }
+  inline virtual void GetLatestDsBlockI(const Json::Value& request,
+                                        Json::Value& response) {
+    (void)request;
+    response = this->GetLatestDsBlock();
+  }
+  inline virtual void GetLatestTxBlockI(const Json::Value& request,
+                                        Json::Value& response) {
+    (void)request;
+    response = this->GetLatestTxBlock();
+  }
+  inline virtual void GetBalanceI(const Json::Value& request,
+                                  Json::Value& response) {
+    response = this->GetBalance(request[0u].asString());
+  }
+  inline virtual void GetMinimumGasPriceI(const Json::Value& request,
+                                          Json::Value& response) {
+    (void)request;
+    response = this->GetMinimumGasPrice();
+  }
+  inline virtual void GetPrevDSDifficultyI(const Json::Value& request,
+                                           Json::Value& response) {
+    (void)request;
+    response = this->GetPrevDSDifficulty();
+  }
+  inline virtual void GetPrevDifficultyI(const Json::Value& request,
+                                         Json::Value& response) {
+    (void)request;
+    response = this->GetPrevDifficulty();
+  }
+  inline virtual void GetSmartContractsI(const Json::Value& request,
+                                         Json::Value& response) {
+    response = this->GetSmartContracts(request[0u].asString());
+  }
+  inline virtual void GetContractAddressFromTransactionIDI(
+      const Json::Value& request, Json::Value& response) {
+    response =
+        this->GetContractAddressFromTransactionID(request[0u].asString());
+  }
+  inline virtual void GetNumPeersI(const Json::Value& request,
+                                   Json::Value& response) {
+    (void)request;
+    response = this->GetNumPeers();
+  }
+  inline virtual void GetNumTxBlocksI(const Json::Value& request,
+                                      Json::Value& response) {
+    (void)request;
+    response = this->GetNumTxBlocks();
+  }
+  inline virtual void GetNumDSBlocksI(const Json::Value& request,
+                                      Json::Value& response) {
+    (void)request;
+    response = this->GetNumDSBlocks();
+  }
+  inline virtual void GetNumTransactionsI(const Json::Value& request,
+                                          Json::Value& response) {
+    (void)request;
+    response = this->GetNumTransactions();
+  }
+  inline virtual void GetTransactionRateI(const Json::Value& request,
+                                          Json::Value& response) {
+    (void)request;
+    response = this->GetTransactionRate();
+  }
+  inline virtual void GetTxBlockRateI(const Json::Value& request,
+                                      Json::Value& response) {
+    (void)request;
+    response = this->GetTxBlockRate();
+  }
+  inline virtual void GetDSBlockRateI(const Json::Value& request,
+                                      Json::Value& response) {
+    (void)request;
+    response = this->GetDSBlockRate();
+  }
+  inline virtual void DSBlockListingI(const Json::Value& request,
+                                      Json::Value& response) {
+    (void)request;
+    response = this->DSBlockListing(request[0u].asUInt());
+  }
+  inline virtual void TxBlockListingI(const Json::Value& request,
+                                      Json::Value& response) {
+    (void)request;
+    response = this->TxBlockListing(request[0u].asUInt());
+  }
+  inline virtual void GetBlockchainInfoI(const Json::Value& request,
+                                         Json::Value& response) {
+    (void)request;
+    response = this->GetBlockchainInfo();
+  }
+  inline virtual void GetRecentTransactionsI(const Json::Value& request,
+                                             Json::Value& response) {
+    (void)request;
+    response = this->GetRecentTransactions();
+  }
+  inline virtual void GetShardingStructureI(const Json::Value& request,
+                                            Json::Value& response) {
+    (void)request;
+    response = this->GetShardingStructure();
+  }
+  inline virtual void GetNumTxnsTxEpochI(const Json::Value& request,
+                                         Json::Value& response) {
+    (void)request;
+    response = this->GetNumTxnsTxEpoch();
+  }
+  inline virtual void GetNumTxnsDSEpochI(const Json::Value& request,
+                                         Json::Value& response) {
+    (void)request;
+    response = this->GetNumTxnsDSEpoch();
+  }
+  inline virtual void GetSmartContractStateI(const Json::Value& request,
+                                             Json::Value& response) {
+    response = this->GetSmartContractState(request[0u].asString());
+  }
+  inline virtual void GetSmartContractCodeI(const Json::Value& request,
+                                            Json::Value& response) {
+    response = this->GetSmartContractCode(request[0u].asString());
+  }
+  inline virtual void GetSmartContractInitI(const Json::Value& request,
+                                            Json::Value& response) {
+    response = this->GetSmartContractInit(request[0u].asString());
+  }
+  inline virtual void GetTransactionsForTxBlockI(const Json::Value& request,
+                                                 Json::Value& response) {
+    response = this->GetTransactionsForTxBlock(request[0u].asString());
+  }
+  inline virtual void GetShardMembersI(const Json::Value& request,
+                                       Json::Value& response) {
+    response = this->GetShardMembers(request[0u].asUInt());
+  }
+
+  virtual std::string GetNetworkId();
+  virtual Json::Value CreateTransaction(const Json::Value& _json);
+  virtual Json::Value GetTransaction(const std::string& transactionHash);
+  virtual Json::Value GetDsBlock(const std::string& blockNum);
+  virtual Json::Value GetTxBlock(const std::string& blockNum);
+  virtual Json::Value GetLatestDsBlock();
+  virtual Json::Value GetLatestTxBlock();
+  virtual Json::Value GetBalance(const std::string& address);
+  virtual std::string GetMinimumGasPrice();
+  virtual Json::Value GetSmartContracts(const std::string& address);
+  virtual std::string GetContractAddressFromTransactionID(
+      const std::string& tranID);
+  virtual unsigned int GetNumPeers();
+  virtual std::string GetNumTxBlocks();
+  virtual std::string GetNumDSBlocks();
+  virtual std::string GetNumTransactions();
+  virtual double GetTransactionRate();
+  virtual double GetTxBlockRate();
+  virtual double GetDSBlockRate();
+  virtual uint8_t GetPrevDSDifficulty();
+  virtual uint8_t GetPrevDifficulty();
+  virtual Json::Value GetShardMembers(unsigned int shardID);
+  virtual Json::Value DSBlockListing(unsigned int page);
+  virtual Json::Value TxBlockListing(unsigned int page);
+  virtual Json::Value GetBlockchainInfo();
+  virtual Json::Value GetRecentTransactions();
+  virtual Json::Value GetShardingStructure();
+  virtual std::string GetNumTxnsDSEpoch();
+  virtual std::string GetNumTxnsTxEpoch();
+
+  size_t GetNumTransactions(uint64_t blockNum);
+  bool ValidateTxn(const Transaction& tx, const Address& fromAddr,
+                   const Account* sender) const;
+  bool StartCollectorThread();
+  std::string GetNodeState();
+
+  static void AddToRecentTransactions(const dev::h256& txhash);
+
+  // gets the number of transaction starting from block blockNum to most recent
+  // block
+
+  Json::Value GetSmartContractState(const std::string& address);
+  Json::Value GetSmartContractInit(const std::string& address);
+  Json::Value GetSmartContractCode(const std::string& address);
+  Json::Value GetTransactionsForTxBlock(const std::string& txBlockNum);
+};
+
+#endif  //__LOOKUP_SERVER_H__
