+++ src/libServer/Server.h
@@ -15,6 +15,9 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
+#ifndef __SERVER_H__
+#define __SERVER_H__
+
 #include "jsonrpccpp/server.h"
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wunused-parameter"
@@ -24,11 +27,18 @@
 #include <random>
 #include "libData/BlockData/BlockHeader/BlockHeaderBase.h"
 #include "libData/DataStructures/CircularArray.h"
+#include "libMediator/Mediator.h"
 
 class Mediator;
 
-class AbstractZServer : public jsonrpc::AbstractServer<AbstractZServer> {
+class Server {
+ protected:
+  Mediator& m_mediator;
+
  public:
+  Server(Mediator& mediator) : m_mediator(mediator) {}
+  ~Server();
+
   enum RPCErrorCode {
     //! Standard JSON-RPC 2.0 errors
     // RPC_INVALID_REQUEST is internally mapped to HTTP_BAD_REQUEST (400).
@@ -66,504 +76,28 @@ class AbstractZServer : public jsonrpc::AbstractServer<AbstractZServer> {
     ERROR
   };
 
-  AbstractZServer(jsonrpc::AbstractServerConnector& conn,
-                  jsonrpc::serverVersion_t type = jsonrpc::JSONRPC_SERVER_V2)
-      : jsonrpc::AbstractServer<AbstractZServer>(conn, type) {
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetNetworkId", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetNetworkIdI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("CreateTransaction", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_OBJECT, NULL),
-        &AbstractZServer::CreateTransactionI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetTransaction", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetTransactionI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetDsBlock", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetDsBlockI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetTxBlock", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetTxBlockI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetLatestDsBlock", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, NULL),
-        &AbstractZServer::GetLatestDsBlockI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetLatestTxBlock", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, NULL),
-        &AbstractZServer::GetLatestTxBlockI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetBalance", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetBalanceI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetMinimumGasPrice", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetMinimumGasPriceI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetPrevDSDifficulty", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_INTEGER, NULL),
-        &AbstractZServer::GetPrevDSDifficultyI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetPrevDifficulty", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_INTEGER, NULL),
-        &AbstractZServer::GetPrevDifficultyI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetSmartContracts", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_ARRAY, "param01", jsonrpc::JSON_STRING,
-                           NULL),
-        &AbstractZServer::GetSmartContractsI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetContractAddressFromTransactionID",
-                           jsonrpc::PARAMS_BY_POSITION, jsonrpc::JSON_STRING,
-                           "param01", jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetContractAddressFromTransactionIDI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetNumPeers", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_INTEGER, NULL),
-        &AbstractZServer::GetNumPeersI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetNumTxBlocks", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetNumTxBlocksI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetNumDSBlocks", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetNumDSBlocksI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetNumTransactions", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetNumTransactionsI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetTransactionRate", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_REAL, NULL),
-        &AbstractZServer::GetTransactionRateI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetTxBlockRate", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_REAL, NULL),
-        &AbstractZServer::GetTxBlockRateI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetDSBlockRate", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_REAL, NULL),
-        &AbstractZServer::GetDSBlockRateI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetCurrentMiniEpoch", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetCurrentMiniEpochI);
-    this->bindAndAddMethod(jsonrpc::Procedure("GetLatestEpochStatesUpdated",
-                                              jsonrpc::PARAMS_BY_POSITION,
-                                              jsonrpc::JSON_STRING, NULL),
-                           &AbstractZServer::GetLatestEpochStatesUpdatedI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetCurrentDSEpoch", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetCurrentDSEpochI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("DSBlockListing", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_INTEGER, NULL),
-        &AbstractZServer::DSBlockListingI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("TxBlockListing", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_INTEGER, NULL),
-        &AbstractZServer::TxBlockListingI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetBlockchainInfo", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, NULL),
-        &AbstractZServer::GetBlockchainInfoI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetRecentTransactions", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, NULL),
-        &AbstractZServer::GetRecentTransactionsI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetShardingStructure", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, NULL),
-        &AbstractZServer::GetShardingStructureI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetNumTxnsTxEpoch", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetNumTxnsTxEpochI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetNumTxnsDSEpoch", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetNumTxnsDSEpochI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetSmartContractState", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetSmartContractStateI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetSmartContractCode", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetSmartContractCodeI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetSmartContractInit", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetSmartContractInitI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetTransactionsForTxBlock",
-                           jsonrpc::PARAMS_BY_POSITION, jsonrpc::JSON_OBJECT,
-                           "param01", jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetTransactionsForTxBlockI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetNodeType", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetNodeTypeI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetDSCommittee", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, NULL),
-        &AbstractZServer::GetDSCommitteeI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetNodeState", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::GetNodeStateI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("IsTxnInMemPool", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::IsTxnInMemPoolI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("GetShardMembers", jsonrpc::PARAMS_BY_POSITION,
-                           jsonrpc::JSON_OBJECT, "param01",
-                           jsonrpc::JSON_INTEGER, NULL),
-        &AbstractZServer::GetShardMembersI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("AddToBlacklistExclusion",
-                           jsonrpc::PARAMS_BY_POSITION, jsonrpc::JSON_BOOLEAN,
-                           "param01", jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::AddToBlacklistExclusionI);
-    this->bindAndAddMethod(
-        jsonrpc::Procedure("RemoveFromBlacklistExclusion",
-                           jsonrpc::PARAMS_BY_POSITION, jsonrpc::JSON_BOOLEAN,
-                           "param01", jsonrpc::JSON_STRING, NULL),
-        &AbstractZServer::RemoveFromBlacklistExclusionI);
-  }
-
-  inline virtual void GetNetworkIdI(const Json::Value& request,
-                                    Json::Value& response) {
-    (void)request;
-    response = this->GetNetworkId();
-  }
-  inline virtual void CreateTransactionI(const Json::Value& request,
-                                         Json::Value& response) {
-    response = this->CreateTransaction(request[0u]);
-  }
-  inline virtual void GetTransactionI(const Json::Value& request,
-                                      Json::Value& response) {
-    response = this->GetTransaction(request[0u].asString());
-  }
-  inline virtual void GetDsBlockI(const Json::Value& request,
-                                  Json::Value& response) {
-    response = this->GetDsBlock(request[0u].asString());
-  }
-  inline virtual void GetTxBlockI(const Json::Value& request,
-                                  Json::Value& response) {
-    response = this->GetTxBlock(request[0u].asString());
-  }
-  inline virtual void GetLatestDsBlockI(const Json::Value& request,
-                                        Json::Value& response) {
-    (void)request;
-    response = this->GetLatestDsBlock();
-  }
-  inline virtual void GetLatestTxBlockI(const Json::Value& request,
-                                        Json::Value& response) {
-    (void)request;
-    response = this->GetLatestTxBlock();
-  }
-  inline virtual void GetBalanceI(const Json::Value& request,
-                                  Json::Value& response) {
-    response = this->GetBalance(request[0u].asString());
-  }
-  inline virtual void GetMinimumGasPriceI(const Json::Value& request,
-                                          Json::Value& response) {
-    (void)request;
-    response = this->GetMinimumGasPrice();
-  }
-  inline virtual void GetPrevDSDifficultyI(const Json::Value& request,
-                                           Json::Value& response) {
-    (void)request;
-    response = this->GetPrevDSDifficulty();
-  }
-  inline virtual void GetPrevDifficultyI(const Json::Value& request,
-                                         Json::Value& response) {
-    (void)request;
-    response = this->GetPrevDifficulty();
-  }
-  inline virtual void GetSmartContractsI(const Json::Value& request,
-                                         Json::Value& response) {
-    response = this->GetSmartContracts(request[0u].asString());
-  }
-  inline virtual void GetContractAddressFromTransactionIDI(
-      const Json::Value& request, Json::Value& response) {
-    response =
-        this->GetContractAddressFromTransactionID(request[0u].asString());
-  }
-  inline virtual void GetNumPeersI(const Json::Value& request,
-                                   Json::Value& response) {
-    (void)request;
-    response = this->GetNumPeers();
-  }
-  inline virtual void GetNumTxBlocksI(const Json::Value& request,
-                                      Json::Value& response) {
-    (void)request;
-    response = this->GetNumTxBlocks();
-  }
-  inline virtual void GetNumDSBlocksI(const Json::Value& request,
-                                      Json::Value& response) {
-    (void)request;
-    response = this->GetNumDSBlocks();
-  }
-  inline virtual void GetNumTransactionsI(const Json::Value& request,
-                                          Json::Value& response) {
-    (void)request;
-    response = this->GetNumTransactions();
-  }
-  inline virtual void GetTransactionRateI(const Json::Value& request,
-                                          Json::Value& response) {
-    (void)request;
-    response = this->GetTransactionRate();
-  }
-  inline virtual void GetTxBlockRateI(const Json::Value& request,
-                                      Json::Value& response) {
-    (void)request;
-    response = this->GetTxBlockRate();
-  }
-  inline virtual void GetDSBlockRateI(const Json::Value& request,
-                                      Json::Value& response) {
-    (void)request;
-    response = this->GetDSBlockRate();
-  }
   inline virtual void GetCurrentMiniEpochI(const Json::Value& request,
                                            Json::Value& response) {
     (void)request;
     response = this->GetCurrentMiniEpoch();
   }
-  inline virtual void GetLatestEpochStatesUpdatedI(const Json::Value& request,
-                                                   Json::Value& response) {
-    (void)request;
-    response = this->GetLatestEpochStatesUpdated();
-  }
+
   inline virtual void GetCurrentDSEpochI(const Json::Value& request,
                                          Json::Value& response) {
     (void)request;
     response = this->GetCurrentDSEpoch();
   }
-  inline virtual void DSBlockListingI(const Json::Value& request,
-                                      Json::Value& response) {
-    (void)request;
-    response = this->DSBlockListing(request[0u].asUInt());
-  }
-  inline virtual void TxBlockListingI(const Json::Value& request,
-                                      Json::Value& response) {
-    (void)request;
-    response = this->TxBlockListing(request[0u].asUInt());
-  }
-  inline virtual void GetBlockchainInfoI(const Json::Value& request,
-                                         Json::Value& response) {
-    (void)request;
-    response = this->GetBlockchainInfo();
-  }
-  inline virtual void GetRecentTransactionsI(const Json::Value& request,
-                                             Json::Value& response) {
-    (void)request;
-    response = this->GetRecentTransactions();
-  }
-  inline virtual void GetShardingStructureI(const Json::Value& request,
-                                            Json::Value& response) {
-    (void)request;
-    response = this->GetShardingStructure();
-  }
-  inline virtual void GetNumTxnsTxEpochI(const Json::Value& request,
-                                         Json::Value& response) {
-    (void)request;
-    response = this->GetNumTxnsTxEpoch();
-  }
-  inline virtual void GetNumTxnsDSEpochI(const Json::Value& request,
-                                         Json::Value& response) {
-    (void)request;
-    response = this->GetNumTxnsDSEpoch();
-  }
-  inline virtual void GetSmartContractStateI(const Json::Value& request,
-                                             Json::Value& response) {
-    response = this->GetSmartContractState(request[0u].asString());
-  }
-  inline virtual void GetSmartContractCodeI(const Json::Value& request,
-                                            Json::Value& response) {
-    response = this->GetSmartContractCode(request[0u].asString());
-  }
-  inline virtual void GetSmartContractInitI(const Json::Value& request,
-                                            Json::Value& response) {
-    response = this->GetSmartContractInit(request[0u].asString());
-  }
-  inline virtual void GetTransactionsForTxBlockI(const Json::Value& request,
-                                                 Json::Value& response) {
-    response = this->GetTransactionsForTxBlock(request[0u].asString());
-  }
+
   inline virtual void GetNodeTypeI(const Json::Value& request,
                                    Json::Value& response) {
     (void)request;
     response = this->GetNodeType();
   }
-  inline virtual void GetDSCommitteeI(const Json::Value& request,
-                                      Json::Value& response) {
-    (void)request;
-    response = this->GetDSCommittee();
-  }
-  inline virtual void GetNodeStateI(const Json::Value& request,
-                                    Json::Value& response) {
-    (void)request;
-    response = this->GetNodeState();
-  }
 
-  inline virtual void IsTxnInMemPoolI(const Json::Value& request,
-                                      Json::Value& response) {
-    response = this->IsTxnInMemPool(request[0u].asString());
-  }
-  inline virtual void GetShardMembersI(const Json::Value& request,
-                                       Json::Value& response) {
-    response = this->GetShardMembers(request[0u].asUInt());
-  }
-  inline virtual void AddToBlacklistExclusionI(const Json::Value& request,
-                                               Json::Value& response) {
-    response = this->AddToBlacklistExclusion(request[0u].asString());
-  }
-  inline virtual void RemoveFromBlacklistExclusionI(const Json::Value& request,
-                                                    Json::Value& response) {
-    response = this->RemoveFromBlacklistExclusion(request[0u].asString());
-  }
-  virtual std::string GetNetworkId() = 0;
-  virtual Json::Value CreateTransaction(const Json::Value& param01) = 0;
-  virtual Json::Value GetTransaction(const std::string& param01) = 0;
-  virtual Json::Value GetDsBlock(const std::string& param01) = 0;
-  virtual Json::Value GetTxBlock(const std::string& param01) = 0;
-  virtual Json::Value GetLatestDsBlock() = 0;
-  virtual Json::Value GetLatestTxBlock() = 0;
-  virtual Json::Value GetBalance(const std::string& param01) = 0;
-  virtual std::string GetMinimumGasPrice() = 0;
-  virtual Json::Value GetSmartContracts(const std::string& param01) = 0;
-  virtual std::string GetContractAddressFromTransactionID(
-      const std::string& param01) = 0;
-  virtual unsigned int GetNumPeers() = 0;
-  virtual std::string GetNumTxBlocks() = 0;
-  virtual std::string GetNumDSBlocks() = 0;
-  virtual std::string GetNumTransactions() = 0;
-  virtual double GetTransactionRate() = 0;
-  virtual double GetTxBlockRate() = 0;
-  virtual double GetDSBlockRate() = 0;
-  virtual uint8_t GetPrevDSDifficulty() = 0;
-  virtual uint8_t GetPrevDifficulty() = 0;
-  virtual std::string GetCurrentMiniEpoch() = 0;
-  virtual std::string GetLatestEpochStatesUpdated() = 0;
-  virtual std::string GetCurrentDSEpoch() = 0;
-  virtual Json::Value DSBlockListing(unsigned int param01) = 0;
-  virtual Json::Value TxBlockListing(unsigned int param01) = 0;
-  virtual Json::Value GetBlockchainInfo() = 0;
-  virtual Json::Value GetRecentTransactions() = 0;
-  virtual Json::Value GetShardingStructure() = 0;
-  virtual std::string GetNumTxnsDSEpoch() = 0;
-  virtual std::string GetNumTxnsTxEpoch() = 0;
-  virtual Json::Value GetSmartContractState(const std::string& param01) = 0;
-  virtual Json::Value GetSmartContractInit(const std::string& param01) = 0;
-  virtual Json::Value GetSmartContractCode(const std::string& param01) = 0;
-  virtual Json::Value GetTransactionsForTxBlock(const std::string& param01) = 0;
-  virtual std::string GetNodeType() = 0;
-  virtual Json::Value GetDSCommittee() = 0;
-  virtual std::string GetNodeState() = 0;
-
-  virtual Json::Value IsTxnInMemPool(const std::string& param01) = 0;
-  virtual Json::Value GetShardMembers(unsigned int param01) = 0;
-  virtual bool AddToBlacklistExclusion(const std::string& ipAddr) = 0;
-  virtual bool RemoveFromBlacklistExclusion(const std::string& ipAddr) = 0;
-};
-
-class Server : public AbstractZServer {
-  Mediator& m_mediator;
-  // Each function of this library can exist in a seperate thread
-
-  std::mutex m_mutexBlockTxPair;
-  std::pair<uint64_t, boost::multiprecision::uint128_t> m_BlockTxPair;
-  std::mutex m_mutexTxBlockCountSumPair;
-  std::pair<uint64_t, boost::multiprecision::uint128_t> m_TxBlockCountSumPair;
-  uint64_t m_StartTimeTx;
-  uint64_t m_StartTimeDs;
-  std::mutex m_mutexDSBlockCache;
-  std::pair<uint64_t, CircularArray<std::string>> m_DSBlockCache;
-  std::mutex m_mutexTxBlockCache;
-  std::pair<uint64_t, CircularArray<std::string>> m_TxBlockCache;
-  static CircularArray<std::string> m_RecentTransactions;
-  static std::mutex m_mutexRecentTxns;
-  std::mt19937 m_eng;
-
- public:
-  Server(Mediator& mediator, jsonrpc::AbstractServerConnector& server);
-  ~Server();
-
-  virtual std::string GetNetworkId();
-  virtual Json::Value CreateTransaction(const Json::Value& _json);
-  virtual Json::Value GetTransaction(const std::string& transactionHash);
-  virtual Json::Value GetDsBlock(const std::string& blockNum);
-  virtual Json::Value GetTxBlock(const std::string& blockNum);
-  virtual Json::Value GetLatestDsBlock();
-  virtual Json::Value GetLatestTxBlock();
-  virtual Json::Value GetBalance(const std::string& address);
-  virtual std::string GetMinimumGasPrice();
-  virtual Json::Value GetSmartContracts(const std::string& address);
-  virtual std::string GetContractAddressFromTransactionID(
-      const std::string& tranID);
-  virtual unsigned int GetNumPeers();
-  virtual std::string GetNumTxBlocks();
-  virtual std::string GetNumDSBlocks();
-  virtual std::string GetNumTransactions();
-  virtual double GetTransactionRate();
-  virtual double GetTxBlockRate();
-  virtual double GetDSBlockRate();
-  virtual uint8_t GetPrevDSDifficulty();
-  virtual uint8_t GetPrevDifficulty();
   virtual std::string GetCurrentMiniEpoch();
-  virtual std::string GetLatestEpochStatesUpdated();
   virtual std::string GetCurrentDSEpoch();
-  virtual Json::Value DSBlockListing(unsigned int page);
-  virtual Json::Value TxBlockListing(unsigned int page);
-  virtual Json::Value GetBlockchainInfo();
-  virtual Json::Value GetRecentTransactions();
-  virtual Json::Value GetShardingStructure();
-  virtual std::string GetNumTxnsDSEpoch();
-  virtual std::string GetNumTxnsTxEpoch();
   virtual std::string GetNodeType();
-  virtual Json::Value GetDSCommittee();
-
-  virtual Json::Value IsTxnInMemPool(const std::string& tranID);
-  virtual Json::Value GetShardMembers(unsigned int shardID);
-
-  virtual bool AddToBlacklistExclusion(const std::string& ipAddr);
-  virtual bool RemoveFromBlacklistExclusion(const std::string& ipAddr);
-
-  static void AddToRecentTransactions(const dev::h256& txhash);
-
-  // gets the number of transaction starting from block blockNum to most recent
-  // block
-  size_t GetNumTransactions(uint64_t blockNum);
-  bool ValidateTxn(const Transaction& tx, const Address& fromAddr,
-                   const Account* sender) const;
   ContractType GetTransactionType(const Transaction& tx) const;
-  bool StartCollectorThread();
-  std::string GetNodeState();
-
-  Json::Value GetSmartContractState(const std::string& address);
-  Json::Value GetSmartContractInit(const std::string& address);
-  Json::Value GetSmartContractCode(const std::string& address);
-  Json::Value GetTransactionsForTxBlock(const std::string& txBlockNum);
 };
+
+#endif  //__SERVER_H__
