+++ src/libZilliqa/Zilliqa.cpp
@@ -158,15 +158,17 @@ Zilliqa::Zilliqa(const PairOfKey& key, const Peer& peer, SyncType syncType,
   m_validator = make_shared<Validator>(m_mediator);
 
   if (LOOKUP_NODE_MODE) {
-    m_serverConnector = make_unique<SafeHttpServer>(RPC_PORT);
-
-  } else {
-    m_serverConnector = make_unique<SafeTcpSocketServer>(IP_TO_BIND, RPC_PORT);
+    m_lookupServerConnector = make_unique<SafeHttpServer>(RPC_PORT);
+    m_lookupServer =
+        make_unique<LookupServer>(m_mediator, *m_lookupServerConnector);
   }
-  if (m_serverConnector == nullptr) {
-    LOG_GENERAL(FATAL, "m_serverConnector NULL");
+  m_statusServerConnector =
+      make_unique<SafeTcpSocketServer>(IP_TO_BIND, RPC_PORT + 1);
+  m_statusServer =
+      make_unique<StatusServer>(m_mediator, *m_statusServerConnector);
+  if (m_statusServer == nullptr) {
+    LOG_GENERAL(FATAL, "m_statusServer NULL");
   }
-  m_server = make_unique<Server>(m_mediator, *m_serverConnector);
 
   m_mediator.RegisterColleagues(&m_ds, &m_n, &m_lookup, m_validator.get());
 
@@ -181,7 +183,7 @@ Zilliqa::Zilliqa(const PairOfKey& key, const Peer& peer, SyncType syncType,
   if (ARCHIVAL_LOOKUP && !LOOKUP_NODE_MODE) {
     LOG_GENERAL(FATAL, "Archvial lookup is true but not lookup ");
   } else if (ARCHIVAL_LOOKUP && LOOKUP_NODE_MODE) {
-    m_server->StartCollectorThread();
+    m_lookupServer->StartCollectorThread();
   }
 
   P2PComm::GetInstance().SetSelfPeer(peer);
@@ -333,15 +335,24 @@ Zilliqa::Zilliqa(const PairOfKey& key, const Peer& peer, SyncType syncType,
       m_lookup.SetServerTrue();
     }
 
-    if (m_server == nullptr) {
-      LOG_GENERAL(INFO, "Pointer unitialized");
+    if (m_lookupServer == nullptr) {
+      LOG_GENERAL(INFO, "Lookup Server Unitialized");
     } else {
-      if ((LOOKUP_NODE_MODE) || (ENABLE_STATUS_RPC)) {
-        if (m_server->StartListening()) {
+      if (LOOKUP_NODE_MODE) {
+        if (m_lookupServer->StartListening()) {
           LOG_GENERAL(INFO, "API Server started successfully");
         } else {
           LOG_GENERAL(WARNING, "API Server couldn't start");
         }
+      }
+    }
+    if (m_statusServer == nullptr) {
+      LOG_GENERAL(INFO, "Status Server Unitialized");
+    } else if (ENABLE_STATUS_RPC) {
+      if (m_statusServer->StartListening()) {
+        LOG_GENERAL(INFO, "Status Server started successfully");
+      } else {
+        LOG_GENERAL(WARNING, "Status Server couldn't start");
       }
     }
   };
